requirements.txt
mss~=9.0.1
pillow~=10.4.0
pygame~=2.6.0
numpy~=2.0.0